requirements.txt
asgiref==3.8.1
Django==5.1.6
dotenv==0.9.9
greenlet==3.1.1
numpy==2.2.3
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
